# swagger-idpay.yaml
openapi: 3.0.3
info:
  title: IDPAY ITN IO API
  version: '1.0'
  contact:
    name: PagoPA S.p.A.
    email: cstar@pagopa.it
  description: |-
    IDPAY ITN IBAN IO

    IDPAY ITN Onboarding Workflow IO

    IDPAY ITN Timeline IO

    IDPAY ITN Wallet IO

    IDPAY ITN Payment QRCODE IO

    IDPAY ITN Payment IO
servers:
  - description: Development Test
    url: 'https://api-io.dev.cstar.pagopa.it/idpay-itn/wallet'
    x-internal: true
  - description: User Acceptance Test
    url: 'https://api-io.uat.cstar.pagopa.it/idpay-itn/wallet'
    x-internal: true
security:
  - bearerAuth: []
...
paths:
...
  /wallet/:
    get:
      tags:
        - wallet
      summary: >-
        ENG: Returns the list of active initiatives of a citizen - IT: Ritorna
        la lista di iniziative attive di un cittadino
      description: Returns the list of active initiatives of a citizen
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          description: 'ENG: Language - IT: Lingua'
          in: header
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletDTO'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_GENERIC_ERROR
                message: Application error
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
  '/wallet/{initiativeId}/detail':
    get:
      tags:
        - wallet
      summary: >-
        ENG: Returns the detail of an initiative - IT: Ritorna i dettagli
        dell'iniziativa
      description: Returns the detail of an initiative
      operationId: getInitiativeBeneficiaryDetail
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: initiativeId
          in: path
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '$ ^[a-zA-Z0-9]+$'
        - name: Accept-Language
          description: 'ENG: Language - IT: Lingua'
          in: header
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeDetailDTO'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '404':
          description: The requested inititative was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_NOT_FOUND
                message: Initiative not found
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeErrorDTO'
              example:
                code: INITIATIVE_GENERIC_ERROR
                message: Application error
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
  '/wallet/{initiativeId}':
    get:
      tags:
        - wallet
      summary: >-
        ENG: Returns the detail of an active initiative of a citizen - IT:
        Ritorna i dettagli di una iniziativa di un cittadino
      description: Returns the detail of an active initiative of a citizen
      operationId: getWalletDetail
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          description: 'ENG: Language - IT: Lingua'
          in: header
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
        - name: initiativeId
          in: path
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '$ ^[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InitiativeDTO1'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '404':
          description: The requested resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_USER_NOT_ONBOARDED
                message: User not onboarded on this initiative
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_GENERIC_ERROR
                message: Application error
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
  '/wallet/{initiativeId}/iban':
    put:
      tags:
        - wallet
      summary: >-
        ENG: Association of an IBAN to an initiative - IT: Associa un IBAN ad
        un'iniziativa
      description: Association of an IBAN to an initiative
      operationId: enrollIban
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          description: 'ENG: Language - IT: Lingua'
          in: header
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
        - name: initiativeId
          in: path
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '$ ^[a-zA-Z0-9]+$'
      requestBody:
        description: >-
          ENG: Unique identifier of the subscribed initiative, IBAN of the
          citizen - IT: Identificativo dell'iniziativa sottoscritta, IBAN del
          cittadino
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IbanPutDTO'
      responses:
        '200':
          description: Enrollment OK
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_ENROLL_IBAN_NOT_ALLOWED_FOR_DISCOUNT_INITIATIVE
                message: >-
                  It is not possible enroll an iban for a discount type
                  initiative
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_GENERIC_ERROR
                message: Application error
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
...
  '/wallet/{initiativeId}/instruments':
    get:
      tags:
        - wallet
      summary: >-
        ENG: Returns the list of payment instruments associated to the
        initiative by the citizen - IT: Ritorna la lista di istrumenti di
        pagamenti associati ad un'iniziativa del cittadino
      description: >-
        Returns the list of payment instruments associated to the initiative by
        the citizen
      operationId: getInstrumentList
      parameters:
        - $ref: '#/components/parameters/ApiVersionHeader'
        - name: Accept-Language
          in: header
          description: 'ENG: Language - IT: Lingua'
          schema:
            type: string
            pattern: '^[ -~]{2,5}$'
            minLength: 2
            maxLength: 5
            example: it-IT
            default: it-IT
          required: true
        - name: initiativeId
          in: path
          description: 'ENG: The initiative ID - IT: Identificativo dell''iniziativa'
          required: true
          schema:
            type: string
            maxLength: 24
            pattern: '$ ^[a-zA-Z0-9]+$'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstrumentListDTO'
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletErrorDTO'
              example:
                code: WALLET_INVALID_REQUEST
                message: Something went wrong handling the request
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '401':
          description: Authentication failed
          content:
            application/json: {}
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '429':
          description: Too many Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInstrumentErrorDTO'
              example:
                code: PAYMENT_INSTRUMENT_TOO_MANY_REQUESTS
                message: Too many requests
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
        '500':
          description: Server ERROR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentInstrumentErrorDTO'
              example:
                code: PAYMENT_INSTRUMENT_GENERIC_ERROR
                message: Something gone wrong while send RTD instrument notify
          headers:
            Access-Control-Allow-Origin:
              $ref: '#/components/headers/Access-Control-Allow-Origin'
            RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
            Retry-After:
              $ref: '#/components/headers/Retry-After'
...
components:
  schemas:
...
    InitiativeErrorDTO:
      type: object
      properties:
        code:
          type: string
          enum:
            - INITIATIVE_INVALID_LOCALE_FORMAT
            - INITIATIVE_INVALID_REQUEST
            - INITIATIVE_NOT_FOUND
            - INITIATIVE_TOO_MANY_REQUESTS
            - INITIATIVE_GENERIC_ERROR
          description: |-
            'ENG: Error code:
             INITIATIVE_INVALID_LOCALE_FORMAT: Initiative not found due to invalid locale format,
             INITIATIVE_INVALID_REQUEST: Something went wrong handling the request,
             INITIATIVE_NOT_FOUND: Initiative not found,
             INITIATIVE_TOO_MANY_REQUESTS: Too many requests,
             INITIATIVE_GENERIC_ERROR: Application error,
             - IT: Codice di errore:
             INITIATIVE_INVALID_LOCALE_FORMAT: Iniziativa non trovata a causa di format locale non valido,
             INITIATIVE_INVALID_REQUEST: Qualcosa è andato storto durante l'invio della richiesta,
             INITIATIVE_NOT_FOUND: Iniziativa non trovata,
             INITIATIVE_TOO_MANY_REQUESTS: Troppe richieste,
             INITIATIVE_GENERIC_ERROR: Errore generico'
        message:
          type: string
          description: 'ENG: Error message - IT: Messaggio di errore'
...
    IbanPutDTO:
      title: IbanPutDTO
      type: object
      required:
        - iban
        - description
      properties:
        iban:
          type: string
          description: 'ENG: IBAN of the citizen - IT: IBAN del cittadino'
          pattern: '^IT[0-9]{2}[A-Z]{1}[0-9]{5}[0-9]{5}[A-Z0-9]{12}$'
          minLength: 27
          maxLength: 27
        description:
          type: string
          description: >-
            ENG: Further information about the iban - IT: Ulteriori informazioni
            sull'iban
          pattern: '^[a-zA-Z0-9_]+$'
          maxLength: 255
...
    WalletDTO:
      type: object
      required:
        - initiativeList
      properties:
        initiativeList:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/InitiativeDTO1'
          description: >-
            ENG: The list of active initiatives of a citizen - IT: Lista delle
            iniziative attive di un cittadino
    InstrumentListDTO:
      type: object
      required:
        - instrumentList
      properties:
        instrumentList:
          type: array
          maxItems: 20
          items:
            $ref: '#/components/schemas/InstrumentDTO'
          description: >-
            'ENG: The list of payment instruments associated to the initiative
            by the citizen - IT: Lista degli strumenti di pagamenti di un
            cittadino associati ad una iniziativa'
...
  parameters:
    ApiVersionHeader:
      name: X-Api-Version
      in: header
      description: 'ENG: Api Version - IT: Versione dell Api'
      required: true
      schema:
        type: string
        enum:
          - v1
        example: v1
        default: v1
  headers:
    Access-Control-Allow-Origin:
      description: >-
        Indicates whether the response can be shared with requesting code from
        the given origin
      schema:
        type: string
        pattern: '^[ -~]{1,2048}$'
        minLength: 1
        maxLength: 2048
    RateLimit-Limit:
      description: The number of allowed requests in the current period
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 240
    RateLimit-Reset:
      description: The number of seconds left in the current period
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 60
    Retry-After:
      description: The number of seconds to wait before allowing a follow-up request
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 240
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer